requirements.txt
Flask
Flask-SocketIO
Flask-Cors
mysql-connector-python
python-socketio
